feat: move decodeInputData to helpers
