+++ tasks/dev/safe.ts
@@ -3,7 +3,6 @@ import {FORK} from "../../helpers/hardhat-constants";
 import {ethers} from "ethers";
 import {decodeMulti} from "ethers-multisend";
 import {SafeTransactionDataPartial} from "@safe-global/safe-core-sdk-types";
-import InputDataDecoder from "ethereum-input-data-decoder";
 import Safe from "@safe-global/safe-core-sdk";
 import EthersAdapter from "@safe-global/safe-ethers-lib";
 import SafeServiceClient from "@safe-global/safe-service-client";
@@ -28,38 +27,18 @@ const TIME_LOCK_SIGS = {
   "0x4dd18bf5": "setPendingAdmin(address)",
 };
 
-const getAbi = async () => {
-  const {
-    ACLManager__factory,
-    ExecutorWithTimelock__factory,
-    ParaSpaceOracle__factory,
-    PoolAddressesProvider__factory,
-    PoolConfigurator__factory,
-    ReservesSetupHelper__factory,
-  } = await import("../../types");
-
-  return [
-    ...ReservesSetupHelper__factory.abi,
-    ...ExecutorWithTimelock__factory.abi,
-    ...PoolAddressesProvider__factory.abi,
-    ...PoolConfigurator__factory.abi,
-    ...ParaSpaceOracle__factory.abi,
-    ...ACLManager__factory.abi,
-  ];
-};
-
 task("decode-safe-txs", "Decode safe txs").setAction(async (_, DRE) => {
   await DRE.run("set-DRE");
   const {getFirstSigner, getTimeLockExecutor} = await import(
     "../../helpers/contracts-getters"
   );
+  const {decodeInputData} = await import("../../helpers/contracts-helpers");
   const timeLock = await getTimeLockExecutor();
   const signer = await getFirstSigner();
   const ethAdapter = new EthersAdapter({
     ethers,
     signerOrProvider: signer,
   });
-  const decoder = new InputDataDecoder(await getAbi());
 
   const safeService = new SafeServiceClient({
     txServiceUrl: `https://safe-transaction-${
@@ -113,11 +92,7 @@ task("decode-safe-txs", "Decode safe txs").setAction(async (_, DRE) => {
     const {to, data} = tx;
     console.log();
     console.log(to);
-    const inputData = decoder.decodeData(data);
-    const normalized = JSON.stringify(inputData, (k, v) => {
-      return v.type === "BigNumber" ? +v.hex.toString(10) : v;
-    });
-    console.log(JSON.stringify(JSON.parse(normalized), null, 4));
+    await decodeInputData(data);
   }
 });
