+++ tasks/dev/timeLock.ts
@@ -2,27 +2,6 @@ import {BigNumber} from "ethers";
 import {task} from "hardhat/config";
 import {DRY_RUN, GLOBAL_OVERRIDES} from "../../helpers/hardhat-constants";
 import {waitForTx} from "../../helpers/misc-utils";
-import InputDataDecoder from "ethereum-input-data-decoder";
-
-const getAbi = async () => {
-  const {
-    ACLManager__factory,
-    ExecutorWithTimelock__factory,
-    ParaSpaceOracle__factory,
-    PoolAddressesProvider__factory,
-    PoolConfigurator__factory,
-    ReservesSetupHelper__factory,
-  } = await import("../../types");
-
-  return [
-    ...ReservesSetupHelper__factory.abi,
-    ...ExecutorWithTimelock__factory.abi,
-    ...PoolAddressesProvider__factory.abi,
-    ...PoolConfigurator__factory.abi,
-    ...ParaSpaceOracle__factory.abi,
-    ...ACLManager__factory.abi,
-  ];
-};
 
 task("next-execution-time", "Next valid execution time").setAction(
   async (_, DRE) => {
@@ -157,20 +136,15 @@ task("decode-queued-txs", "Decode queued transactions").setAction(
     const {getTimeLockExecutor} = await import(
       "../../helpers/contracts-getters"
     );
+    const {decodeInputData} = await import("../../helpers/contracts-helpers");
     const timeLock = await getTimeLockExecutor();
-    const decoder = new InputDataDecoder(await getAbi());
     const filter = timeLock.filters.QueuedAction();
     const events = await timeLock.queryFilter(filter);
     for (const e of events) {
       if (!(await timeLock.isActionQueued(e.args.actionHash))) {
         continue;
       }
-      const inputData = decoder.decodeData(e.args.data.toString());
-      const normalized = JSON.stringify(inputData, (k, v) => {
-        return v.type === "BigNumber" ? +v.hex.toString(10) : v;
-      });
-
-      console.log(JSON.stringify(JSON.parse(normalized), null, 4));
+      decodeInputData(e.args.data.toString());
       console.log();
     }
   }
@@ -182,17 +156,12 @@ task("decode-buffered-txs", "Decode buffered transactions").setAction(
     const {getTimeLockDataInDb} = await import(
       "../../helpers/contracts-helpers"
     );
-    const decoder = new InputDataDecoder(await getAbi());
+    const {decodeInputData} = await import("../../helpers/contracts-helpers");
     const actions = await getTimeLockDataInDb();
 
     for (const a of actions) {
       const [, , , data] = a.action;
-      const inputData = decoder.decodeData(data.toString());
-      const normalized = JSON.stringify(inputData, (k, v) => {
-        return v.type === "BigNumber" ? +v.hex.toString(10) : v;
-      });
-
-      console.log(JSON.stringify(JSON.parse(normalized), null, 4));
+      decodeInputData(data.toString());
       console.log();
     }
   }
